fix mime type issue by renaming jsx to js
