+++ src/components/DailyCheckIn.js
@@ -2,7 +2,7 @@ import React from 'https://cdn.skypack.dev/react';
 import { Card } from './ui/card.js';
 import { Button } from './ui/button.js';
 import { Textarea } from './ui/textarea.js';
-import SectionTitle from './SectionTitle.jsx';
+import SectionTitle from './SectionTitle.js';
 import { useCollection } from '../firebase.js';
 
 export default function DailyCheckIn({ userId }) {
