+++ src/components/ProfileSettings.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'https://cdn.skypack.dev/react';
 import { Mic, Camera as CameraIcon } from 'https://cdn.skypack.dev/lucide-react';
 import { Card } from './ui/card.js';
 import { Textarea } from './ui/textarea.js';
-import SectionTitle from './SectionTitle.jsx';
+import SectionTitle from './SectionTitle.js';
 import { db, getDoc, doc } from '../firebase.js';
 
 export default function ProfileSettings({ userId, ageRange, onChangeAgeRange }) {
